Move lat_long_to_octant's logic to lib
